Move logout functionality to navigation
