+++ src/app/components/MainNav/index.js
@@ -1,88 +1,14 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import { withRouter } from 'react-router'
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
 
-import navData from '../../../data/mainNavigationData'
+import { setAccount } from '../../actions/account'
 
-import bars from '../../../img/bars.svg'
-import style from './MainNav.css'
+import MainNav from './MainNav'
 
-class MainNav extends Component {
-  state = {
-    menuIsOpen: false,
-  }
-
-  componentDidMount() {
-    window.addEventListener('click', this.closeDropdownMenu)
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('click', this.closeDropdownMenu)
-  }
-
-  toggleMenu = event => {
-    this.setState({ menuIsOpen: !this.state.menuIsOpen })
-  }
-
-  closeDropdownMenu = event => {
-    if (!event.target.className.includes('MainNav')) {
-      this.setState(prevState => ({ menuIsOpen: false }))
-    }
-  }
-
-  pushHistory = event => {
-    let path = event.target.dataset.target || event.target.parentNode.dataset.target
-    const { history } = this.props
-
-    console.log(path)
-
-    history.push(path)
-    this.toggleMenu()
-  }
-
-  generateNavigationMarkup = () => {
-    return navData.map(navItem => {
-      return (
-        <button className={ style.mainNavLinkButton } data-target={ navItem.path } key={ navItem.id }>
-          <img src={ navItem.img } alt={ navItem.alt } className={ style.mainNavLinkButtonImg } />
-          {navItem.title}
-        </button>
-      )
-    })
-  }
-
-  getDropdownClasses = () => {
-    const { menuIsOpen } = this.state
-
-    let classes
-    if (menuIsOpen) {
-      classes = `${style.mainNavDropdown} ${style.mainNavDropdownOpen}`
-    } else {
-      classes = style.mainNavDropdown
-    }
-
-    return classes
-  }
-
-  render() {
-    const navigationMarkup = this.generateNavigationMarkup()
-    const dropdownMenuClasses = this.getDropdownClasses()
-
-    return (
-      <nav className={ style.mainNav }>
-        <button className={ style.mainNavButton } onClick={ this.toggleMenu }>
-          <img src={ bars } alt='three vertical bars' className={ style.mainNavButtonImg } />
-        </button>
-        <div className={ dropdownMenuClasses } onClick={ this.pushHistory }>
-          {navigationMarkup}
-        </div>
-      </nav>
-    )
-  }
+const actionCreators = {
+  setAccount,
 }
 
-export default withRouter(MainNav)
+const mapDispatchToProps = dispatch => bindActionCreators(actionCreators, dispatch)
 
-MainNav.propTypes = {
-  history: PropTypes.object,
-}
+export default connect(null, mapDispatchToProps)(MainNav)
