+++ src/app/containers/Home/RenameAccount.js
@@ -6,16 +6,13 @@ import PrimaryButton from '../../components/common/buttons/PrimaryButton'
 import saveSVG from '../../../img/save.svg'
 import style from './Home.css'
 
-const RenameAccount = ({ accountName, onClickHandler, onChangeHandler }) => (
+const RenameAccount = ({ accountName, onSubmitHandler, onChangeHandler }) => (
   <div className={ style.renameAccountContainer }>
     <h1 className={ style.renameAccountHeading }>Rename account</h1>
-    <InputField onChangeHandler={ onChangeHandler } value={ accountName } />
-    <PrimaryButton
-      buttonText='Save'
-      icon={ saveSVG }
-      classNames={ style.renameAccountButton }
-      onClickHandler={ onClickHandler }
-    />
+    <form onSubmit={ onSubmitHandler }>
+      <InputField onChangeHandler={ onChangeHandler } value={ accountName } />
+      <PrimaryButton buttonText='Save' icon={ saveSVG } classNames={ style.renameAccountButton } />
+    </form>
   </div>
 )
